+++ Solution/Wiki-CUDA-LSA/ParallelTFIDF/tfidf.cuh
@@ -2,6 +2,6 @@
 #ifndef _INC_TFIDF // include guard for 3rd party interop
 #define _INC_TFIDF
 
-void tfidf(int rows, int cols, double * tmatrix);
+void tfidf(int rows, int cols, float * tmatrix);
 
 #endif // _INC_TDIDF
